python-pipeline/requirements.txt
# Alternative requirements with flexible versions for Python 3.11+
# Use this if you encounter version conflicts with requirements.txt

# Core FastAPI and web framework
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.12,<1.0.0
aiofiles>=24.0.0,<25.0.0

# Core Data Science Stack - Compatible versions
numpy>=1.24.0,<2.0.0  # Using NumPy 1.x for better compatibility
scipy>=1.10.0,<1.13.0  # Compatible with NumPy 1.x and gensim
scikit-learn>=1.3.0,<1.6.0  # Compatible with NumPy 1.x
pandas>=2.0.0,<3.0.0

# Machine Learning and NLP
sentence-transformers>=3.0.0,<4.0.0
torch>=2.0.0,<3.0.0
transformers>=4.35.0,<5.0.0

# NLP Libraries with compatible versions
spacy>=3.7.0,<4.0.0
gensim>=4.3.2,<5.0.0  # This version works with NumPy 1.x
textblob>=0.18.0,<1.0.0
nltk>=3.8.0,<4.0.0

# Text Processing and Similarity
fuzzywuzzy>=0.18.0,<1.0.0
python-Levenshtein>=0.25.0,<1.0.0

# AI and Language Models
google-generativeai>=0.7.0,<1.0.0

# Visualization and Analysis (optional)
plotly>=5.20.0,<6.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
wordcloud>=1.9.0,<2.0.0

# HTTP Requests and utilities
requests>=2.31.0,<3.0.0
click>=8.1.0,<9.0.0

# Additional dependencies that may be needed
joblib>=1.3.0,<2.0.0
threadpoolctl>=3.1.0,<4.0.0
packaging>=23.0,<25.0
regex>=2023.0.0,<2026.0.0
tqdm>=4.65.0,<5.0.0

# Optional: For better tokenization support
tokenizers>=0.15.0,<1.0.0

# Development and testing (optional)
# pytest>=7.4.0,<8.0.0
# pytest-asyncio>=0.21.0,<1.0.0

# RSS and web scraping
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.8.5

# HTML cleaning and parsing  
html5lib==1.1
bleach==6.0.0

# Date parsing
python-dateutil==2.8.2

# MongoDB driver
pymongo>=4.0.0